requirements.txt
# Web Scraping
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2

# Google Calendar Integration
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Date/Time Processing
python-dateutil==2.8.2
pytz==2023.3

# Scheduling
schedule==1.2.0

# Configuration
python-dotenv==1.0.0

# Utilities
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
loguru==0.7.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0